refactor: restructure booking service into modular architecture
