+++ apps/api/src/services/booking/cron/booking-cron-manager.ts
